extract day actions into separate component
